+++ src/test/java/nextstep/subway/line/domain/PathFinderTest.java
@@ -1,8 +1,5 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.line.domain;
 
-import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.Lines;
-import nextstep.subway.line.domain.Section;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.Stations;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,7 +11,6 @@ import java.util.NoSuchElementException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class PathFinderTest {
